+++ sdk/typespec/typespec_client_core/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 0.5.0 (Unreleased)
+## 0.5.0 (2025-07-10)
 
 ### Features Added
 
@@ -13,8 +13,6 @@ the default asynchronous runtime with another.
 - Minimum supported Rust version (MSRV) is now 1.85.
 - Removed `Pipeline::replace_policy`.
 - Renamed `typespec_client_core::date` to `typespec_client_core::time` and added `typespec_client_core::time::Duration`
-
-### Bugs Fixed
 
 ### Other Changes